MnistExample/requirements.txt
pandas
boto3
Pillow
sagemaker-inference
tensorflow==2.17.0
keras==3.5.0
flask==2.0.1
gunicorn==23.0.0
Werkzeug==2.2.2